BQL Query Planning
query_plan() takes a query object (parsed BQL statement) and a mounted
database, and attempts to resolve metric / tag references to series
btree indices, in both the SELECT clause and the (optional) WHERE
clause.  This aids in execution by figuring out which series are in play
for a given query.