Implicit Aggregation
If an AGGREGATE clause exists in a BQL query, but a field lacks an
aggregation function, the evaluator will now apply an implicit CF to the
final (pre-return) value of the field.

This introduces some new BQL AGGREGATE sub-clauses, so that the same
flexibility afforded at the bucketing level (number of samples, CF used)
is available at the aggregation level.