Stop runaway memory usage in the agent
New fixed-size send buffer ensures that if an agent is unable to relay
its results up to the core, it doesn't keep allocating memory in a
growing backlog, ultimately OOMing the system and possible endangering
other operations-focused process.  Instead, the agent will begin
ejecting the oldest results, in an attempt to stay up-to-date when
relaying resumes.